requirements.txt
python-docx>=1.1.0
openai>=1.12.0
tiktoken>=0.6.0
click>=8.1.7
python-dotenv>=1.0.0
tqdm>=4.66.0
